+++ noodles-gff/CHANGELOG.md
@@ -36,7 +36,8 @@
 
     This splits the line into its key-optional value components.
 
-  * gff/record: Parse phase (`Record::phase`).
+  * gff/record: Parse score (`Record::score`) and phase
+    (`Record::phase`).
 
 ### Removed